improve invite/uninvite
invitation and uninvitations are send continuously.
invitation stops if we get a format message or we time out (source timeout). in the first case, we send an 'add' event; in the second case we send a 'invite timeout' event.
uninvitation stops once the source times out, in which case the we send a regular 'remove' event. note that this might never happen, if the source decides to keep sending!
in pd/sc clients, we make sure that we check for the given sink to filter redundant 'invite'/'uninvite' events.