pd/aoo_send~: automatically update format
we store the original format and re-parse with the appropriate default values
everytime DSP is updated and the blocksize, channel count or sample rate has changed.

this way users can simply use a format with default values (e.g. just "pcm") and have it updated automatically.
This is particularly relevant for multichannel mode because the input channel count can change dynamically!